Refactor container handlers into separate files
